+++ platform-tests/src/test/java/org/junit/platform/launcher/core/DefaultLauncherTests.java
@@ -18,8 +18,6 @@ import static org.junit.platform.commons.util.CollectionUtils.getOnlyElement;
 import static org.junit.platform.engine.TestExecutionResult.successful;
 import static org.junit.platform.engine.discovery.DiscoverySelectors.selectPackage;
 import static org.junit.platform.engine.discovery.DiscoverySelectors.selectUniqueId;
-import static org.junit.platform.launcher.EngineFilter.excludeEngines;
-import static org.junit.platform.launcher.EngineFilter.includeEngines;
 import static org.junit.platform.launcher.core.LauncherDiscoveryRequestBuilder.DEFAULT_DISCOVERY_LISTENER_CONFIGURATION_PROPERTY_NAME;
 import static org.junit.platform.launcher.core.LauncherDiscoveryRequestBuilder.request;
 import static org.junit.platform.launcher.core.LauncherFactoryForTestingPurposesOnly.createLauncher;
@@ -367,117 +365,6 @@ class DefaultLauncherTests {
 		assertThat(testPlan.getRoots()).hasSize(2);
 		assertThat(testPlan.getChildren(UniqueId.forEngine("engine1").toString())).hasSize(1);
 		assertThat(testPlan.getChildren(UniqueId.forEngine("engine2").toString())).hasSize(1);
-	}
-
-	@Test
-	void launcherWillNotExecuteEnginesIfNotIncludedByAnEngineFilter() {
-		var firstEngine = new DemoHierarchicalTestEngine("first");
-		TestDescriptor test1 = firstEngine.addTest("test1", noOp);
-		var secondEngine = new DemoHierarchicalTestEngine("second");
-		TestDescriptor test2 = secondEngine.addTest("test2", noOp);
-
-		var launcher = createLauncher(firstEngine, secondEngine);
-
-		// @formatter:off
-		var testPlan = launcher.discover(
-			request()
-				.selectors(selectUniqueId(test1.getUniqueId()), selectUniqueId(test2.getUniqueId()))
-				.filters(includeEngines("first"))
-				.build());
-		// @formatter:on
-
-		assertThat(testPlan.getRoots()).hasSize(1);
-		var rootIdentifier = testPlan.getRoots().iterator().next();
-		assertThat(testPlan.getChildren(rootIdentifier.getUniqueId())).hasSize(1);
-		assertThat(testPlan.getChildren(UniqueId.forEngine("first").toString())).hasSize(1);
-	}
-
-	@Test
-	void launcherWillExecuteAllEnginesExplicitlyIncludedViaSingleEngineFilter() {
-		var firstEngine = new DemoHierarchicalTestEngine("first");
-		TestDescriptor test1 = firstEngine.addTest("test1", noOp);
-		var secondEngine = new DemoHierarchicalTestEngine("second");
-		TestDescriptor test2 = secondEngine.addTest("test2", noOp);
-
-		var launcher = createLauncher(firstEngine, secondEngine);
-
-		// @formatter:off
-		var testPlan = launcher.discover(
-			request()
-				.selectors(selectUniqueId(test1.getUniqueId()), selectUniqueId(test2.getUniqueId()))
-				.filters(includeEngines("first", "second"))
-				.build());
-		// @formatter:on
-
-		assertThat(testPlan.getRoots()).hasSize(2);
-	}
-
-	@Test
-	void launcherWillNotExecuteEnginesExplicitlyIncludedViaMultipleCompetingEngineFilters() {
-		var firstEngine = new DemoHierarchicalTestEngine("first");
-		TestDescriptor test1 = firstEngine.addTest("test1", noOp);
-		var secondEngine = new DemoHierarchicalTestEngine("second");
-		TestDescriptor test2 = secondEngine.addTest("test2", noOp);
-
-		var launcher = createLauncher(firstEngine, secondEngine);
-
-		// @formatter:off
-		var testPlan = launcher.discover(
-			request()
-				.selectors(selectUniqueId(test1.getUniqueId()), selectUniqueId(test2.getUniqueId()))
-				.filters(includeEngines("first"), includeEngines("second"))
-				.build());
-		// @formatter:on
-
-		assertThat(testPlan.getRoots()).isEmpty();
-	}
-
-	@Test
-	void launcherWillNotExecuteEnginesExplicitlyExcludedByAnEngineFilter() {
-		var firstEngine = new DemoHierarchicalTestEngine("first");
-		TestDescriptor test1 = firstEngine.addTest("test1", noOp);
-		var secondEngine = new DemoHierarchicalTestEngine("second");
-		TestDescriptor test2 = secondEngine.addTest("test2", noOp);
-
-		var launcher = createLauncher(firstEngine, secondEngine);
-
-		// @formatter:off
-		var testPlan = launcher.discover(
-			request()
-				.selectors(selectUniqueId(test1.getUniqueId()), selectUniqueId(test2.getUniqueId()))
-				.filters(excludeEngines("second"))
-				.build());
-		// @formatter:on
-
-		assertThat(testPlan.getRoots()).hasSize(1);
-		var rootIdentifier = testPlan.getRoots().iterator().next();
-		assertThat(testPlan.getChildren(rootIdentifier.getUniqueId())).hasSize(1);
-		assertThat(testPlan.getChildren(UniqueId.forEngine("first").toString())).hasSize(1);
-	}
-
-	@Test
-	void launcherWillExecuteEnginesHonoringBothIncludeAndExcludeEngineFilters() {
-		var firstEngine = new DemoHierarchicalTestEngine("first");
-		TestDescriptor test1 = firstEngine.addTest("test1", noOp);
-		var secondEngine = new DemoHierarchicalTestEngine("second");
-		TestDescriptor test2 = secondEngine.addTest("test2", noOp);
-		var thirdEngine = new DemoHierarchicalTestEngine("third");
-		TestDescriptor test3 = thirdEngine.addTest("test3", noOp);
-
-		var launcher = createLauncher(firstEngine, secondEngine, thirdEngine);
-
-		// @formatter:off
-		var testPlan = launcher.discover(
-			request()
-				.selectors(selectUniqueId(test1.getUniqueId()), selectUniqueId(test2.getUniqueId()), selectUniqueId(test3.getUniqueId()))
-				.filters(includeEngines("first", "second"), excludeEngines("second"))
-				.build());
-		// @formatter:on
-
-		assertThat(testPlan.getRoots()).hasSize(1);
-		var rootIdentifier = testPlan.getRoots().iterator().next();
-		assertThat(testPlan.getChildren(rootIdentifier.getUniqueId())).hasSize(1);
-		assertThat(testPlan.getChildren(UniqueId.forEngine("first").toString())).hasSize(1);
 	}
 
 	@Test
